moved heart beat logic to separate file
